requirements.txt
absl-py==1.4.0
addict==2.4.0
argcomplete==3.0.8
asttokens==2.2.1
av2==0.1.0
backcall==0.2.0
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
descartes==1.1.0
efficientnet-pytorch==0.7.1
einops==0.6.0
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
future==0.18.3
GeometricKernelAttention==1.0
google-auth==2.17.2
google-auth-oauthlib==0.4.6
grpcio==1.54.0rc1
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
llvmlite==0.31.0
lyft-dataset-sdk==0.0.8
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv-full==1.4.0
mmdet==2.14.0
mmsegmentation==0.14.1
networkx==2.2
nox==2023.4.22
numba==0.48.0
numpy==1.19.5
nuscenes-devkit==1.1.10
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.4.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
plyfile==0.8.1
prettytable==3.6.0
prompt-toolkit==3.0.38
protobuf==3.19.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
Pygments==2.14.0
pyparsing==3.0.9
pyproj==3.5.0
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
Shapely==1.8.0
similaritymeasures==0.7.0
six==1.16.0
stack-data==0.6.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.12
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.65.0
traitlets==5.9.0
trimesh==2.35.39
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.23.0
wcwidth==0.2.6
Werkzeug==2.2.3
yapf==0.32.0
zipp==3.15.0
